ArjunReddy AI Smart Virtual Assistant/requirements.txt
speech_recognition==3.8.1
pyttsx3==2.90
pyowm==3.4.0
webbrowser
random2
newsapi-python==0.2.6
wikipedia-api==0.5.4
wikipedia==1.4.0
spacy==3.0.6
